Added InGame Menu
- Added menu
- moved transitions to Transition Manager
- Added subclass for all UI Menus
- Made AudioManager check for other instances of itself
