+++ Assets/AudioManager.cs
@@ -6,8 +6,12 @@ using Pixelplacement;
 public class AudioManager : Singleton<AudioManager> {
 
     // Use this for initialization
-    void Start () {
-		
+    void Awake () {
+        AudioManager instance = FindObjectOfType<AudioManager>();
+        if(instance != this)
+        {
+            Destroy(gameObject);
+        }
 	}
 	
 	// Update is called once per frame
