+++ Assets/InGameMenu.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 6c8bf7571df35594aa3d3882fc34ee31
+timeCreated: 1516565483
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
