+++ Assets/SCRIPTS/TitleScreen/MenuUI.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 40d6620aad768274bb3e7bb432e7637a
+timeCreated: 1516564769
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
