+++ Assets/TransitionManager.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: d29fe3afcc84cc44a8180e17d989c527
+timeCreated: 1516563485
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
